% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Leslie.R
\name{create.Leslie}
\alias{create.Leslie}
\title{Create a Leslie matrix
Currently coded with pre-determined parameters, lambda set to be 1.02}
\usage{
create.Leslie(M, F)
}
\arguments{
\item{M}{natural mortality rate}

\item{F}{fishing mortality rate}
}
\description{
Create a Leslie matrix
Currently coded with pre-determined parameters, lambda set to be 1.02
}
\examples{
create.Leslie
}
