% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.transient.R
\name{iterate.transient}
\alias{iterate.transient}
\title{Function to iterate an age-structured matrix model before and after MPA implementation
Leslie matrix LF (with fishing), L0 (without)}
\usage{
iterate.transient(LF, L0)
}
\arguments{
\item{LF}{Leslie matrix with fishing}

\item{L0}{Leslie matrix without fishing}
}
\description{
Function to iterate an age-structured matrix model before and after MPA implementation
Leslie matrix LF (with fishing), L0 (without)
}
\examples{
iterate.transient
}
