% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA.transient.R
\name{MPA.transient}
\alias{MPA.transient}
\title{Transient analysis of post-MPA dynamics}
\usage{
MPA.transient(M, F)
}
\arguments{
\item{M}{natural mortality rate}

\item{F}{fishing mortality rate}
}
\description{
Uses a pre-determined age-structured matrix model
Inputs: natural mortality rate, M; pre-MPA fishing rate, F
}
\examples{
MPA.transient
}
