% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.Leslie.R
\name{optim.Leslie}
\alias{optim.Leslie}
\title{find the value of fecundity in a Leslie matrix that gives a particular lambda1 value}
\usage{
optim.Leslie(x, L, Target)
}
\arguments{
\item{x}{multiplier (to be modified by optim)}

\item{L}{Leslie matrix}

\item{Target}{target eigenvalue}
}
\description{
find the value of fecundity in a Leslie matrix that gives a particular lambda1 value
}
