% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.results.R
\name{plot.results}
\alias{plot.results}
\title{Plot the results of the matrix model}
\usage{
\method{plot}{results}(D, L)
}
\arguments{
\item{D}{results object from iterate.model}

\item{L}{Leslie matrix}
}
\description{
Plot the results of the matrix model
}
\examples{
plot.results
}
